create model to referentielijsten config
